pull in late init code from mux_stats_late_init
